pyproject.toml
[tool.poetry]
name = "tt-backend"
version = "1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^4.2.6"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
cython = "^3.0.3"
torch = "^2.1.0"
scipy = "^1.11.3"
numpy = "^1.26.0"
torchvision = "^0.16.0"
unidecode = "^1.3.7"
torchtext = "^0.16.0"
transformers = "^4.34.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
